+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 - Calendar Widget to visualize repetitions left for months/days (heatmap coloration, plugged to backend)
 - Plug timezone when querying reps calendar
+- Reschedule widget to change current node due date
 
 ### Fixed
 - In node tree, left triangle was grayed out even if the current node had a non-dismissed fragment child, when that fragment only had spore children. It is not the case anymore.